1.6.1 + Exit
* Minified the packaged monaco-editor and removed a lot of unused dependencies

* Bump to 1.6.1

* Add meta tags for description and viewports

* Make 'blue' contrast a lot more for accessibility and to comply with WCAG AA(A)

Most likely will be the final change to this project, I've graduated and no longer have access to a parallax propeller. It was nice working on this project and I hope it reaches students that had the same feeling about the block-programming as me.

Hopefully I can make something for arduino or the raspberry pi next. :)